# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/task": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Create task",
                "parameters": [
                    {
                        "description": "Request 新增 Task (status 預設為 0)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apireq.CreateTask"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apires.CreateTask"
                        }
                    },
                    "400": {
                        "description": "{\"code\":\"400400\",\"message\":\"Wrong parameter format or invalid\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    },
                    "500": {
                        "description": "{\"code\":\"500002\",\"message\":\"Database insert error\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    }
                }
            }
        },
        "/task/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "task_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request 修改 Task",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apireq.UpdateTask"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apires.UpdateTask"
                        }
                    },
                    "400": {
                        "description": "{\"code\":\"400400\",\"message\":\"Wrong parameter format or invalid\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    },
                    "404": {
                        "description": "{\"code\":\"400404\",\"message\":\"Resource not found\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    },
                    "500": {
                        "description": "{\"code\":\"500003\",\"message\":\"Database update error\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "task_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"code\":\"400400\",\"message\":\"Wrong parameter format or invalid\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    },
                    "404": {
                        "description": "{\"code\":\"400404\",\"message\":\"Resource not found\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    },
                    "500": {
                        "description": "{\"code\":\"500004\",\"message\":\"Database delete error\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "List task",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apires.ListTask"
                        }
                    },
                    "500": {
                        "description": "{\"code\":\"500001\",\"message\":\"Database query error\"}",
                        "schema": {
                            "$ref": "#/definitions/errors.AppErrorMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apireq.CreateTask": {
            "type": "object",
            "required": [
                "name",
                "status"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
        "apireq.UpdateTask": {
            "type": "object",
            "required": [
                "name",
                "status"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                }
            }
        },
        "apires.CreateTask": {
            "type": "object",
            "properties": {
                "result": {
                    "$ref": "#/definitions/apires.Task"
                }
            }
        },
        "apires.ListTask": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apires.Task"
                    }
                }
            }
        },
        "apires.Task": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "apires.UpdateTask": {
            "type": "object",
            "properties": {
                "result": {
                    "$ref": "#/definitions/apires.Task"
                }
            }
        },
        "errors.AppErrorMsg": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}